sagas/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Faker==22.5.1
fastavro==1.9.4
Flask==3.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
fastapi==0.109.0
httpx==0.26.0
uvicorn==0.27.0.post1
itsdangerous==2.1.2
Jinja2==3.1.3
localStoragePy==0.2.3
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
packaging==23.2
psycopg2-binary==2.9.5
pulsar-client==3.4.0
PyDispatcher==2.0.7
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
